+++ semTools/NEWS.md
@@ -7,6 +7,7 @@
 
 ## Bug Fixes:
 
+- `emmeans` methods deal with models fit with missing data.
 - `net()` returned an error for models with categorical indicators, now fixed.
 - The wrong *df* were used to find confidence limits for RMSEA calculated from the scaled $\chi^2$ statistic in `lavaan.mi` objects.